+++ src/main/java/com/info2022/app/repository/IOrganizationDao.java
@@ -8,6 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import com.info2022.app.dto.OrganizationDto;
 import com.info2022.app.entity.Organization;
 
 @Repository
@@ -15,6 +16,6 @@ public interface IOrganizationDao extends JpaRepository<Organization, Long>{
 
 	@Query("select u from organizations u where u.name = ?1")
 	public List<Organization> findByname(String name);
-	public Organization findByCuit(Integer cuit);
+	public Organization findByCuit(Long cuit);
 	public Optional<Organization> findById(Long id);
 }
